refactor: Rename AtomeInputView to TextFieldView.swift and fix border and placeholder colors
