+++ OmiseSDK/CreditCardFormViewModel.swift
@@ -1,0 +1,9 @@
+//
+//  CreditCardFormViewModel.swift
+//  OmiseSDK
+//
+//  Created by Andrei Solovev on 19/6/23.
+//  Copyright © 2023 Omise. All rights reserved.
+//
+
+import Foundation
